requirements.txt
tencentcloud-sdk-python-dnspod==3.0.1024
requests==2.31.0
python-dotenv==1.0.0
